requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
mysql-connector-python==8.0.33
Werkzeug==2.3.3
joblib==1.3.2
scikit-learn==1.3.0